[cxx-interop] Allow many specializations of a class template
An old piece of logic in ClangImporter was trying to limit the number of instantiations for each C++ class template to prevent long compile times. Unfortunately this started causing hard-to-reproduce deserialization errors on large projects which use many different instantiations of `std::vector` and `std::allocator`.

The instantiation limit was arbitrary, it serves no real purpose and causes issues. This change removes it.

rdar://158397914