build.ps1: accommodate the split SPM runtime build
Adjust the build to accommodate the split SPM manifest runtime build. In
order to support building the toolchain without the just built compiler,
we need to split out the SPM runtime which is used by the new toolchain.
We must build the SPM manifest runtime against the just built runtime as
this is the compiler consuming the runtime.