[tools] Create a new script called git-clang-format-all.zsh
This is a small zsh script that automates running git-clang-format + git rebase
squash on a branch. The intent is that one can just work on ones branch, run
this before creating a PR on github and move on with life. The assumption is
that ones tree is clean without any changes so we should be able to apply
git-clang-format cleanly by applying it to each commit one by one working
backwards from ToT to the passed in commit's parent commit.