[rbi] Clean up isolation history pushing.
Specifically, I made it so that we in code always performed all pushes for an
operation as much at the same time as possible. This is in case I want to move
from a sequence based approach to a single node/operation approach that may make
it easier to emit more specific diagnostics.

I also fixed a small bug in Partition::singleRegion where when we should have
only merged a single function argument, we instead merged in all function
arguments we had already seen as well.