[Frontend] Check diagnostics in generated sources
The previous commit added support for checking diagnostics in other
buffers, enabling diagnostic verification for e.g. macro expansions.
Because the diagnostics are in a different buffer, there is no error for
not marking them as "expected". This commit expands the scope to
include generated sources originating in code in any checked buffer, to
help make sure that errors aren't accidentally missed.