[CS] Avoid introducing Void closure result binding for hole
If we have a hole binding for the result of a closure, avoid
introducing Void as an additional binding since that'll just cause
local solution ambiguities. This doesn't affect regular type-checking
since we end up treating such ambiguities as "identical", so we just
pick one. But for completion, we don't do such filtering so need to
make sure we end up with the hole solution to avoid treating Void as
the contextual type for cases such as:

```
_ = {
    x.#^CC^#
}
```

Currently we avoid this by not handling holes in `repairFailures`
for closure body locators, but I'm changing that in the next commit.