[git-clang-format-all] Apply fixup commits one at a time.
I discovered while using this locally that it is possible to get a series of
rebased commits that cause the rebase to failure. The problem goes away when one
applies the changes one at a time rather than all at once. This is also just a
more robust way to do this since we are not relying on git-clang-format's hunk
logic and git's rebase hunk logic to be in sync.