[Distributed] Allow disabling SerializationRequirement by using Any
It should be possible to disable SerializationRequirement by assigning
Any to it. This effectively disables the checking, since any type
conforms to that.

Previously we would crash trying to check for this conformance; or
rather, emitting diagnostics about un-necessary casts then encountering
incomplete types in CSDiagnostics.

Avoiding the cast entirely sounds like a good solution here.

resolves rdar://159285863

[Distributed] Simplify checking for Any type

Harden diagnoseForcedCastExpr for lack of cast-to type