[unittests] Fix compilation of Runtime unittests
PR#78516 performed a rename from `swiftRuntime` to `swiftRuntimeCore`
but might have skipped these two instances because the code is rarely
compiled in Darwin platforms (and it is compiled in
Linux/Windows platforms either). The code on the runtime unittests seems
to be associated with the SDK overlays being built (either static or
dynamic), which none are for Darwin platforms. Other platforms should
build the dynamic SDK overlay, but the file is surrounded with checks
for `__APPLE__` and `__MACH__`.

The normal CI would never hit this file, but someone might while
building the static or dynamic SDK overlay if they so request it to the
`build-script`.